requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Google APIs
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Google Cloud
google-cloud-secret-manager==2.17.0
google-cloud-firestore==2.14.0
google-cloud-error-reporting==1.1.0

# Template rendering
jinja2==3.1.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1

# HTTP client
httpx==0.25.2

# Google Cloud Tasks
google-cloud-tasks==2.14.2
